+++ .vscode/settings.json
@@ -4,5 +4,8 @@
     },
     "[javascriptreact]": {
         "editor.tabSize": 2
+    },
+    "[scss]": {
+        "editor.tabSize": 2
     }
 }
