--- rework/pages/_utils.tsx
@@ -1,9 +1,0 @@
-import { useEffect, useRef } from "react";
-
-export function usePrevious<T>(value: T): T | undefined {
-  const ref = useRef<T>();
-  useEffect(() => {
-    ref.current = value;
-  }, [value]);
-  return ref.current;
-}
