+++ rework/public/static/favicon-white.ico
