+++ rework/src/pages/index.tsx
@@ -1,0 +1,83 @@
+import Head from 'next/head'
+import { GetServerSidePropsContext, GetServerSidePropsResult } from 'next/types';
+
+import React, { useContext } from 'react';
+
+import { InnerMoon } from '@theme-toggles/react'
+
+import faviconBlack from '../../public/static/favicon-black.ico';
+import faviconWhite from '../../public/static/favicon-white.ico';
+
+import { COOKIE_NAME_THEME, SetThemeContext, Theme, ThemeComponent, ThemeContext } from './_theme';
+import CharacterSheet from '../components/CharacterSheet';
+
+
+type Props = {
+  initialTheme?: Theme;
+}
+
+export default function Home(props: Props) {
+  return (
+    <ThemeComponent initialTheme={props.initialTheme}>
+      <div>
+        <AppHead />
+        <main>
+          <ThemeToggleButton />
+          <CharacterSheet />
+        </main>
+      </div>
+    </ThemeComponent>
+  )
+}
+
+function AppHead() {
+  var theme = useContext(ThemeContext);
+  return (
+    <Head>
+      <title>Character Creator - Arcus</title>
+      <link rel='shortcut icon' href={theme == Theme.Light ? faviconBlack.src : faviconWhite.src} />
+    </Head>
+  )
+}
+
+export function getServerSideProps(context: GetServerSidePropsContext): GetServerSidePropsResult<Props> {
+  const initialThemeName = context.req.cookies[COOKIE_NAME_THEME];
+  const initialTheme: Theme | undefined = initialThemeName ? Theme[initialThemeName] : undefined;
+  const props: Props = {};
+  if (initialTheme) {
+    props.initialTheme = initialTheme
+  }
+  return { props };
+};
+
+
+function ThemeToggleButton() {
+  const theme = useContext(ThemeContext)
+  const setTheme = useContext(SetThemeContext)
+
+  return (
+    <>
+      <style jsx global>{`
+        button.theme-toggle.inner-moon svg {
+          position:absolute;
+          left:50%;
+          top:50%;
+          transform:translate(-50%, -50%);
+        }
+        `}</style>
+      <InnerMoon
+        className={`btn shadow rounded-circle ${theme == Theme.Light ? 'btn-dark' : 'btn-light'} inner-moon position-absolute top-0 end-0 z-1`}
+        style={{
+          margin: '1rem',
+          width: '45px',
+          height: '45px',
+          fontSize: 20,
+          padding: 0,
+          backgroundColor: 'var(--bs-btn-bg)',
+        }}
+        duration={750}
+        toggled={theme == Theme.Light}
+        onToggle={(toggled) => setTheme(toggled ? Theme.Light : Theme.Dark)} />
+    </>
+  )
+}
