--- rework/styles/styling.scss
@@ -1,6 +1,0 @@
-
-// :root {
-//     --bs-primary: #0012da;
-//     --bs-body-color: #202020;
-//     --bs-body-bg: #f1f1f1;
-// }
